Refactor TableSection
